move components to compile view and bounding box synchonously
