+++ src/Features/ERDDAP/Map/index.tsx
@@ -6,7 +6,7 @@ import "ol/ol.css"
 import { useRouter } from "next/navigation"
 import GeoJSON from "ol/format/GeoJSON"
 import { fromLonLat, transformExtent } from "ol/proj"
-import React from "react"
+import React, { useState } from "react"
 import { Button } from "reactstrap"
 import { RFeature, RLayerVector, RMap, RPopup, RStyle } from "rlayers"
 import type { RView } from "rlayers/RMap"
@@ -14,17 +14,18 @@ import type { RView } from "rlayers/RMap"
 import { useStatefulView } from "Features/StatefulMap"
 import { colors } from "Shared/colors"
 import { paths } from "Shared/constants"
-import { BoundingBox } from "Shared/regions"
+import { BoundingBox, regionList } from "Shared/regions"
 import { urlPartReplacer } from "Shared/urlParams"
 import { EsriOceanBasemapLayer, EsriOceanReferenceLayer } from "components/Map"
 
 import { aDayAgoRounded } from "Shared/time"
+import { useParams } from "next/navigation"
 import { UsePlatforms } from "../hooks"
 import { PlatformFeature } from "../types"
 
 export interface Props {
   // Bounding box for fitting to a region
-  boundingBox?: BoundingBox
+  boundingBox?: BoundingBox | null
   //  Platform to highlight
   platformId?: string
   // Height to adjust map to match sidebar
@@ -107,15 +108,23 @@ const PlatformLayer = ({ platform, selected, old = false }: PlatformLayerProps) 
 // Initial view to display if one is not otherwise set
 const initial = { center: fromLonLat([-68.5, 43.5]), zoom: 6 }
 
-export const ErddapMapBase: React.FC<BaseProps> = ({
-  platforms,
-  platformId,
-  height,
-  boundingBox,
-  view,
-  setView,
-}: BaseProps) => {
+export const ErddapMapBase: React.FC<BaseProps> = ({ platforms, platformId, height, view, setView }: BaseProps) => {
   const mapRef = React.useRef<RMap>(null)
+  const params: { regionId?: string } = useParams()
+  const [boundingBox, setBoundingBox] = useState<BoundingBox | null>()
+
+  //If params change, set bounding box
+  React.useEffect(() => {
+    if (typeof params.regionId !== "undefined") {
+      const regionId = decodeURIComponent(params.regionId)
+      const region = regionList.find((r) => r.slug === regionId)
+      setBoundingBox(region?.bbox)
+    }
+    if (typeof params.regionId === "undefined") {
+      console.log(view, "HELLO", initial)
+      setView(initial)
+    }
+  }, [params.regionId])
 
   // When the bounding box gets set, zoom to the region
   React.useEffect(() => {
@@ -125,7 +134,7 @@ export const ErddapMapBase: React.FC<BaseProps> = ({
 
       mapRef?.current?.ol.getView().fit(extent)
     }
-  }, [boundingBox, view, platforms])
+  }, [boundingBox, platforms])
 
   // Make sure the height of the map gets updated when jumping
   // from home to platform view
@@ -158,7 +167,7 @@ export const ErddapMapBase: React.FC<BaseProps> = ({
  */
 export const ErddapMap: React.FC<Props> = ({ platformId, boundingBox, height }: Props) => {
   const [view, handleSetView] = useStatefulView()
-
+  console.log(view)
 
   return (
     <UsePlatforms>
@@ -168,7 +177,6 @@ export const ErddapMap: React.FC<Props> = ({ platformId, boundingBox, height }: 
           view={view}
           setView={handleSetView}
           platformId={platformId}
-          boundingBox={boundingBox}
           height={height}
         />
       )}
